Remove headers.h
