+++ src/addrman.h
@@ -6,7 +6,6 @@
 
 #include "netbase.h"
 #include "protocol.h"
-#include "util.h"
 
 
 #include <map>
