+++ src/checkpoints.cpp
@@ -5,8 +5,10 @@
 #include <boost/assign/list_of.hpp> // for 'map_list_of()'
 #include <boost/foreach.hpp>
 
-#include "headers.h"
 #include "checkpoints.h"
+
+#include "main.h"
+#include "uint256.h"
 
 namespace Checkpoints
 {
