+++ src/db.cpp
@@ -3,12 +3,17 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file license.txt or http://www.opensource.org/licenses/mit-license.php.
 
-#include "headers.h"
 #include "db.h"
-#include "net.h"
+#include "util.h"
+#include "main.h"
+#include "wallet.h"
 #include <boost/version.hpp>
 #include <boost/filesystem.hpp>
 #include <boost/filesystem/fstream.hpp>
+
+#ifndef WIN32
+#include "sys/stat.h"
+#endif
 
 using namespace std;
 using namespace boost;
