+++ src/init.h
@@ -5,6 +5,8 @@
 #ifndef BITCOIN_INIT_H
 #define BITCOIN_INIT_H
 
+#include "wallet.h"
+
 extern CWallet* pwalletMain;
 
 void Shutdown(void* parg);
