+++ src/qt/qtipcserver.cpp
@@ -7,7 +7,7 @@
 #include <boost/tokenizer.hpp>
 #include <boost/date_time/posix_time/posix_time.hpp>
 
-#include "headers.h"
+#include "util.h"
 #include "qtipcserver.h"
 
 using namespace boost::interprocess;
